Split the functional tests classes
